- Include query validation in QueryParser (both pattern and such-that clauses) - Add VARIABLE, CONSTANT, PROGLINE to TypeTable.h - Fix signed-unsigned mismatch in Uses.cpp (change int to size_t)

+++ PowerRangerMain/EmptyGeneralTesting/PKB/Uses.cpp
@@ -84,7 +84,7 @@ vector<int> Uses::getUses(TypeTable::SynType t, VARNAME v){	//Select a such that
 	try{
 		INDEX index = varTable->getVarIndex(v);
 		vector<STMTNUM> ans;
-		for(int i=0;i<usesTable.size();i++){
+		for(size_t i=0;i<usesTable.size();i++){
 			vector<STMTNUM> temp = usesTable[i];
 			vector<STMTNUM>::iterator it = temp.begin();
 			for(;it!=temp.end();++it){
